Release 0.2.0-alpha2 (rename 0.2.0-RC2 to alpha2)   * STATIC folder added at root project and settings.py updated * global.js file created to JS folder at STATIC folder in root project * JS function enableAllFields() added to global.js * JS functions remove disabled and hidden attributes removed from project.js and enabledFields in documens.js (replace for enableAllFields) * In custom_functions.py renamed create_objects to create_or_update_object and updated it * global.js added to documents and projects templates * Refactoring views and forms at projects and documents * document.js updated to accept new buttons and inputs' names

+++ makemake/sites/views.py
@@ -12,7 +12,7 @@ from django.db.models import Q
 
 from makemake.sites.models import Site
 from makemake.sites.forms import SiteForm
-from makemake.core.custom_functions import update_object
+from makemake.core.custom_functions import create_or_update_object
 
 from auditlog.context import set_actor
 
@@ -23,11 +23,15 @@ from makemake.core.custom_functions import separar_valores_sem_espaco, is_list_e
 
 def delete(request, pk):
     if request.user.has_perm('global_permissions.entra_em_sites'):
-        project = Site.objects.get(pk=pk)
-        try:
-            register_to_delete = get_object_or_404(Site, pk=pk)
-            register_to_delete.delete()
-        except IntegrityError:
+        if Site.objects.filter(pk=pk).exists():
+            try:
+                with set_actor(request.user):
+                    register_to_delete = get_object_or_404(Site, pk=pk)
+                    register_to_delete.delete()
+            except IntegrityError:
+                message = "The register couldn't be deleted!"
+                messages.info(request, message)
+        else:
             message = "The register couldn't be deleted!"
             messages.info(request, message)
     else:
@@ -40,8 +44,6 @@ def home(request):
     items = Site.objects.all()
     messages.info(request, None)
     return render(request, 'sites/home.html', {'items': items})
-
-
 
 def search(request):
     # Expressão regular
@@ -112,19 +114,18 @@ def new(request):
         form = SiteForm(request.POST or None)
 
         if form.is_valid():
-            a = form.cleaned_data['name']
-            b = form.cleaned_data['place']
-
+            data = {
+                'name': form.cleaned_data['name'],
+                'place': form.cleaned_data['place'],
+            }
             # Logica para gravar instância principal
-            b2 = Site(name=a, place=b, )
             try:
-                with set_actor(request.user):
-                    b2.save()
+                create_or_update_object(request, Site, data)
             except IntegrityError as e:
                 #messages.add_message(request, messages.ERROR, 'There has been an error...')
-                form.add_error('code', 'Code category must be unique!')
-                #context = {'form': form}
-                #return render(request, 'categories/new.html', context)
+                #form.add_error('code', 'Code category must be unique!')
+                context = {'form': form}
+                return render(request, 'categories/new.html', context)
 
     else: # Empty new form
         form = SiteForm()
@@ -135,29 +136,24 @@ def new(request):
 
 
 def edit(request, pk=None):
-    #extra_forms = 1  # You can set the initial number of forms here
-    #ProjectBuildingFormSet = formset_factory(ProjectBuildingForm, extra=extra_forms)
     if request.user.has_perm('global_permissions.entra_em_sites'):
 
         if request.method == 'POST':    # Newly filled form
             form = SiteForm(request.POST or None)
 
             if form.is_valid():
-                a = form.cleaned_data['name']
-                b = form.cleaned_data['place']
-                # Lista de atributos do modelo Project
-                attributes = [
-                    "name", "place", 
-                ]
-                # Lista de valores correspondentes
-                values = [a, b,]
-                update_object(request, Site, pk, attributes, values)
+                data = {
+                    'name': form.cleaned_data['name'],
+                    'place': form.cleaned_data['place'],
+                }
+                create_or_update_object(request, Site, data, pk)
 
         else: # Read data from 
             instance = Site.objects.get(pk=pk)
             form = SiteForm(instance=instance, prefix='edit')
         context = {'form': form}
         return render(request, 'sites/new_or_edit.html', context)
+    
     message = "You cannot have permission to do this!"
     messages.info(request, message)
     return home(request)
